requirements.txt
altgraph==0.17.4
filelock==3.15.4
fsspec==2024.9.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
sympy==1.13.2
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
